Add files via upload
This is the task 1 of team three that is STACK BY QUEUE & QUEUE BY STACK

The conclusion we have come to that is beneficial b/w two approaches within queue by stack is that we have analyse the data first if mo're no. of eateries are to be pushed , then we must make the pop operation costly and vice versa.
similarly for stack by queue we have analyse the no. of enteteries , if no. of enteries are are less , then enequeue operation is costly else dequeue operation is costly