Project 4
Convert a equation to prefix and postfix form.If given equation is prefix or postfix covert it to normal and evaluate the result .
Convert given prefix equation to postfix and vice versa. The code shall work for both characters and integers and can contain brackets in equation if so boadmas rule for equation evaluation must be followed.